requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
aiosqlite==0.19.0
openpyxl==3.1.2
reportlab==4.0.8
python-dotenv==1.0.0
schedule==1.2.0